models/kokkoro/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scipy==1.11.4
numpy==1.24.0
pydub==0.25.1
soundfile==0.12.1
runpod==1.5.0
kokoro-onnx==0.1.0
<<<<<<< HEAD
onnxruntime==1.16.3
=======
onnxruntime-gpu==1.16.3
>>>>>>> bfb632049c2b3072aacded3e7bdbac32b686bf66
